Programación 4/Biblioteca con Celery y KeyDB/celery_app.py
from celery import Celery
import os
from dotenv import load_dotenv

load_dotenv()

celery = Celery(
    __name__,
    broker=os.getenv('CELERY_BROKER_URL'),
    backend=os.getenv('CELERY_RESULT_BACKEND')
)