Programación 4/Biblioteca con Celery y KeyDB/requirements.txt
Flask
Flask-Mail
celery
redis
python-dotenv